Turing codec Version 1.2
Main features added:
	- New SIMD optimisation functions for intra-prediction and
	  quantisation
	- Revised rate-control algorithm for parallel processing
	- Fixes to field-coding and SAO
	- Added frame doubling functioanlity and temporal scalability
	- Added conformance window settings for interlaced material
	- Added fade detection algorithm

+++ turing/sei/user_data_unregistered.h
@@ -113,7 +113,7 @@ template <> struct Write<user_data_unregistered>
     template <class H> static void go(user_data_unregistered fun, H &h)
     {
         StateWriteUserDataUnregistered *stateWriteUserDataUnregistered = h;
-        fun.payloadSize = 16 + (stateWriteUserDataUnregistered->end - stateWriteUserDataUnregistered->p);
+        fun.payloadSize = 16 + static_cast<int>(stateWriteUserDataUnregistered->end - stateWriteUserDataUnregistered->p);
         Syntax<user_data_unregistered>::go(fun, h);
     }
 };
